Refactor to account for recent changes
