+++ types/src/spaces/oklrch.d.ts
@@ -1,3 +1,4 @@
-import ColorSpace from "../space.js";
 declare const _default: ColorSpace;
 export default _default;
+import ColorSpace from "../ColorSpace.js";
+//# sourceMappingURL=oklrch.d.ts.map
